package struktúra refaktor
