+++ PizzaApp/src/main/java/com/github/holiver98/config/ApplicationConfiguration.java
@@ -1,7 +1,13 @@
 package com.github.holiver98.config;
 
-import com.github.holiver98.dao.*;
+import com.github.holiver98.dal.inmemory.*;
 import com.github.holiver98.database.InMemoryDatabase;
+import com.github.holiver98.service.inmemory.InMemoryCartService;
+import com.github.holiver98.service.inmemory.InMemoryPizzaService;
+import com.github.holiver98.service.inmemory.InMemoryRatingService;
+import com.github.holiver98.service.inmemory.InMemoryUserService;
+import com.github.holiver98.service.jpa.JpaRatingService;
+import com.github.holiver98.service.jpa.JpaUserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import com.github.holiver98.service.*;
