+++ PizzaApp/src/main/java/com/github/holiver98/dal/jpa/IIngredientRepository.java
@@ -1,0 +1,9 @@
+package com.github.holiver98.dal.jpa;
+
+import com.github.holiver98.model.Ingredient;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IIngredientRepository extends JpaRepository<Ingredient, String> {
+}
