--- PizzaApp/src/main/java/com/github/holiver98/repository/IUserRepository.java
@@ -1,12 +1,0 @@
-package com.github.holiver98.repository;
-
-import com.github.holiver98.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface IUserRepository extends JpaRepository<User, String> {
-    Optional<User> findByEmailAddress(String emailAddress);
-}
